requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
timm
eva-decord
ftfy
regex
einops
fvcore
iopath
numpy
matplotlib
types-regex
vtk
mayavi
cartopy
numpy
Pillow
diffusers
librosa
datasets
gradio
streamlit
tensorboard
accelerate
transformers
tokenizers<=0.19
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d